Fix RoPE to use bfloat16 instead of float16 for TPU compatibility
- RoPE was casting output to float16 which causes precision issues on TPU
- Changed to bfloat16 for TPU-friendly computation
- Should fix 462% degradation in REUSE layers on TPU

+++ src/MaxText/layers/kascade_layers.py
@@ -76,7 +76,8 @@ def apply_rope(xq: jnp.ndarray, xk: jnp.ndarray, freqs_cis: jnp.ndarray):
     xq_out = jnp.stack([xq_rotated.real, xq_rotated.imag], axis=-1).reshape(xq.shape)
     xk_out = jnp.stack([xk_rotated.real, xk_rotated.imag], axis=-1).reshape(xk.shape)
     
-    return xq_out.astype(jnp.float16), xk_out.astype(jnp.float16)
+    # TPU-friendly: use bfloat16 instead of float16
+    return xq_out.astype(jnp.bfloat16), xk_out.astype(jnp.bfloat16)
 
 class KascadeAnchorAttention(nn.Module):
     """
